First implementation of AssetGroup/Loader Minor fixes about values Added float operator to LinkedVariables

+++ Sprites/Sprite.cs
@@ -18,20 +18,6 @@ namespace FCSG{
             if(this.texture==null){
                 throw new ArgumentException(message: "The texture of the sprite is null");
             }
-
-            // All this stuff is already handled in SpriteBase
-            // if(group!=null){ //Adds the sprite to the group
-            //     this.groups.Add(group);
-            //     group.Add(this);
-            // }
-            // if (groups!=null){
-            //     foreach(ObjectGroup<SpriteObject> spriteGroup in groups){
-            //         this.groups.Add(spriteGroup);
-            //         spriteGroup.Add(this);
-            //     }
-            // }
-
-            // this.draw=true;
         }
         public override void Draw(bool drawMiddle=true){
             BasicDraw(this.spriteBatch,drawMiddle);
@@ -44,7 +30,7 @@ namespace FCSG{
                 if(drawMiddle==true){
                     DrawMiddleTexture();
                 }
-                spriteBatch.Draw(texture, new Rectangle(this.x,this.y,this.width,this.height),null,color,rotation,origin,effects,depth);
+                spriteBatch.Draw(texture, new Rectangle(this.x,this.y,this.width,this.height),null,color,rotation,origin,effects,(float)depth);
             }
         }
     }
